update pre license

+++ builds/android/build.sh
@@ -8,7 +8,7 @@ ANDROID_BUILD_DIR="$(pwd)"
 source ${ANDROID_BUILD_DIR}/android_build_helper.sh
 
 # Choose a C++ standard library implementation from the ndk
-ANDROID_BUILD_CXXSTL="gnustl_shared_48"
+ANDROID_BUILD_CXXSTL="gnustl_shared_49"
 
 # Set up android build environment and set ANDROID_BUILD_OPTS array
 android_build_env
@@ -25,29 +25,40 @@ if [[ $ANDROID_BUILD_CLEAN ]]; then
     rm -rf "${ANDROID_BUILD_PREFIX}"/*
 fi
 
-##
-# Build libsodium from latest master branch
+if [ -z $CURVE ]; then
+    CURVE="--disable-curve"
+    VERIFY="libzmq.so"
+elif [ $CURVE == "libsodium" ]; then
+    CURVE="--with-libsodium=yes"
+    VERIFY="libzmq.so libsodium.so"
+    ##
+    # Build libsodium from latest master branch
 
-(android_build_verify_so "libsodium.so" &> /dev/null) || {
-    rm -rf "${cache}/libsodium"
-    (cd "${cache}" && git clone --depth 1 git://github.com/jedisct1/libsodium.git) || exit 1
-    (cd "${cache}/libsodium" && ./autogen.sh \
-        && ./configure "${ANDROID_BUILD_OPTS[@]}" --disable-soname-versions \
-        && make -j 4 \
-        && make install) || exit 1
-}
+    (android_build_verify_so "libsodium.so" &> /dev/null) || {
+        rm -rf "${cache}/libsodium"
+        (cd "${cache}" && git clone -b stable --depth 1 git://github.com/jedisct1/libsodium.git) || exit 1
+        (cd "${cache}/libsodium" && ./autogen.sh \
+            && ./configure --quiet "${ANDROID_BUILD_OPTS[@]}" --disable-soname-versions \
+            && make -j 4 \
+            && make install) || exit 1
+    }
+elif [ $CURVE == "tweetnacl" ]; then
+    # Default
+    CURVE=""
+    VERIFY="libzmq.so"
+fi
 
 ##
 # Build libzmq from local source
 
 LIBTOOL_EXTRA_LDFLAGS='-avoid-version'
 
-(android_build_verify_so "libzmq.so" "libsodium.so" &> /dev/null) || {
+(android_build_verify_so ${VERIFY} &> /dev/null) || {
     rm -rf "${cache}/libzmq"
     (cp -r ../.. "${cache}/libzmq" && cd "${cache}/libzmq" && make clean)
     
     (cd "${cache}/libzmq" && ./autogen.sh \
-        && ./configure "${ANDROID_BUILD_OPTS[@]}" --without-documentation --with-libsodium=yes \
+        && ./configure --quiet "${ANDROID_BUILD_OPTS[@]}" ${CURVE} --without-docs \
         && make -j 4 \
         && make install) || exit 1
 }
@@ -55,5 +66,5 @@ LIBTOOL_EXTRA_LDFLAGS='-avoid-version'
 ##
 # Verify shared libraries in prefix
 
-android_build_verify_so "libzmq.so" "libsodium.so"
+android_build_verify_so ${VERIFY}
 echo "libzmq android build succeeded"
